airflow_dags/requirements.txt
apache-airflow==2.9.0
joblib==1.5.0
pandas==2.2.3